Creation of the Menu in its own java file for implementation
I put it in it's own file cause we haven't gotten far enough to use some of its functions like save and load ....
However I got resume ands exit to function and you can bring it up and close it with the escape key.
The buttons are only clickable right now, I will be working on keyboard functionality.